+++ app/src/main/java/app/revanced/manager/ui/component/patches/IncompatiblePatchDialog.kt
@@ -1,0 +1,41 @@
+package app.revanced.manager.ui.component.patches
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.WarningAmber
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import app.revanced.manager.R
+
+@OptIn(ExperimentalMaterial3ExpressiveApi::class)
+@Composable
+fun IncompatiblePatchDialog(
+    appVersion: String,
+    compatibleVersions: Set<String>,
+    onDismissRequest: () -> Unit
+) = AlertDialog(
+    icon = {
+        Icon(Icons.Outlined.WarningAmber, null)
+    },
+    onDismissRequest = onDismissRequest,
+    confirmButton = {
+        TextButton(onClick = onDismissRequest, shapes = ButtonDefaults.shapes()) {
+            Text(stringResource(R.string.ok))
+        }
+    },
+    title = { Text(stringResource(R.string.incompatible_patch)) },
+    text = {
+        Text(
+            stringResource(
+                R.string.app_version_not_compatible,
+                appVersion,
+                compatibleVersions.joinToString(", ")
+            )
+        )
+    }
+)